Move SpecDecodeMetadata to spec decoding manager.
